Make parameters more rational
When the ``Job.run`` method is called, it receives the rank of the MPI process.
There are no more global parameters passed to the slaves: everything should be
put into the Job class. A new function ``run_event_loop`` takes care of calling
the correct event loop (either ``run_master`` or ``run_slave``).