Make ``run_event_loop`` return a list of jobs
The event loop returns a list of the jobs. This can be used to save some values
during the execution of each job and have them conveniently handled at the end
of the event loop itself. The example has been modified to show how to use this
feature.

The documentation string for the module has been expanded as well, providing an
explanation of the intended usage of the library.